requirements.txt
streamlit 
fastapi 
uvicorn 
whisper 
googletrans 
ffmpeg-python
